requirements.txt
openai
rich
python-dotenv
pytest
